Acertando a divisão da Calculadora.

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -36,7 +36,7 @@ public class Calculadora {
 	public static int multiplicacao(int a, int b) {
 		return a * b;
 	}
-	public static int divisao(int a, int b) {
+	public static double divisao(double a, double b) {
 		return a / b;
 	}
 }
